requirements.txt
attrs==23.2.0
automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
exceptiongroup==1.2.2
filelock==3.15.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
mysql-connector-python==9.0.0
packaging==24.1
parsel==1.9.1
pluggy==1.5.0
protego==0.3.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydispatcher==2.0.7
pyopenssl==24.2.1
pytest==8.3.2
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
scrapy==2.11.2
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
tldextract==5.1.2
tomli==2.0.1
twisted==24.3.0
typing-extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
zope-interface==6.4.post2
